optimize the manta example
