+++ examples/manta/scripts/constants.js
@@ -1,0 +1,28 @@
+// Testnet
+const L1_TESTNET_CONTRACTS = {
+  StateCommitmentChain: '0x0000000000000000000000000000000000000000',
+  BondManager: '0x0000000000000000000000000000000000000000',
+  CanonicalTransactionChain: '0x0000000000000000000000000000000000000000',
+  AddressManager: '0x0AaeDFF2961D05021832cA093cf9409eDF5ECa8C',
+  L1CrossDomainMessenger: '0x7Ad11bB9216BC9Dc4CBd488D7618CbFD433d1E75',
+  L1StandardBridge: '0x4638aC6b5727a8b9586D3eba5B44Be4b74ED41Fc',
+  OptimismPortal: '0x7FD7eEA37c53ABf356cc80e71144D62CD8aF27d3',
+  L2OutputOracle: '0x8553D4d201ef97F2b76A28F5E543701b25e55B1b',
+};
+
+// Mainnet
+const L1_MAINNET_CONTRACTS = {
+  StateCommitmentChain: "0x0000000000000000000000000000000000000000",
+  BondManager: "0x0000000000000000000000000000000000000000",
+  CanonicalTransactionChain: "0x0000000000000000000000000000000000000000",
+  AddressManager: "0x3Ad319BB4872F8cB75a26Ac30CC4bD2d56b67b05",
+  L1CrossDomainMessenger: "0x635ba609680c55C3bDd0B3627b4c5dB21b13c310",
+  L1StandardBridge: "0x3B95bC951EE0f553ba487327278cAc44f29715E5",
+  OptimismPortal: "0x9168765EE952de7C6f8fC6FaD5Ec209B960b7622",
+  L2OutputOracle: "0x30c789674ad3B458886BBC9abf42EEe19EA05C1D"
+};
+
+module.exports = {
+  L1_TESTNET_CONTRACTS,
+  L1_MAINNET_CONTRACTS,
+};
